Correccion de Estructura de Proyecto
