+++ src/main/java/com/proyecto/application/dto/bus/BusRequestDTO.java
@@ -1,4 +1,4 @@
-package com.proyecto.application.dto;
+package com.proyecto.application.dto.bus;
 
 import lombok.Builder;
 import lombok.Data;
