+++ src/main/java/com/proyecto/application/dto/viaje/ViajeResponseDTO.java
@@ -1,4 +1,4 @@
-package com.proyecto.application.dto;
+package com.proyecto.application.dto.viaje;
 
 import lombok.Builder;
 import lombok.Data;
@@ -7,9 +7,10 @@ import java.time.LocalDate;
 
 @Data
 @Builder
-public class ViajeRequestDTO {
-    private Long idBus;
-    private Long idDestino;
+public class ViajeResponseDTO {
+    private Long idViaje;
+    private String nombreBus;
+    private String nombreDestino;
     private LocalDate fechaSalida;
     private LocalDate fechaLlegada;
     private String incidencias;
