+++ src/main/java/com/proyecto/application/service/BusService.java
@@ -1,7 +1,7 @@
 package com.proyecto.application.service;
 
-import com.proyecto.application.dto.BusRequestDTO;
-import com.proyecto.application.dto.BusResponseDTO;
+import com.proyecto.application.dto.bus.BusRequestDTO;
+import com.proyecto.application.dto.bus.BusResponseDTO;
 
 import java.util.List;
